[REF] Moves utils.py functions to compute.py
Moves xcorr, zscore, and normalize from utils.py to compute.py. These
functions are all more heavily used in computations throughout the
module than they are used as "utilities". Minor change, but makes more
sense that they be colocated with similar functions.
