+++ pyls/compute.py
@@ -2,7 +2,121 @@
 
 import numpy as np
 from sklearn.utils.extmath import randomized_svd
+from sklearn.utils.validation import check_array
 from pyls import utils
+
+
+def xcorr(X, Y, norm=True):
+    """
+    Calculates the cross-covariance matrix of `X` and `Y`
+
+    Parameters
+    ----------
+    X : (S, B) array_like
+        Input matrix, where `S` is samples and `B` is features.
+    Y : (S, T) array_like, optional
+        Input matrix, where `S` is samples and `T` is features.
+
+    Returns
+    -------
+    xprod : (T, B) `numpy.ndarray`
+        Cross-covariance of `X` and `Y`
+    """
+
+    if X.ndim != Y.ndim:
+        raise ValueError('Number of dims of `X` and `Y` must match.')
+    if X.ndim != 2:
+        raise ValueError('`X` and `Y` must each have 2 dims.')
+    if len(X) != len(Y):
+        raise ValueError('The first dim of `X` and `Y` must match.')
+
+    Xn, Yn = zscore(X), zscore(Y)
+    if norm:
+        Xn, Yn = normalize(Xn), normalize(Yn)
+    xprod = (Yn.T @ Xn) / (len(Xn) - 1)
+
+    return xprod
+
+
+def zscore(data, axis=0, ddof=1, comp=None):
+    """
+    Z-scores `X` by subtracting mean and dividing by standard deviation
+
+    Effectively the same as `np.nan_to_num(scipy.stats.zscore(X))` but
+    handles DivideByZero without issuing annoying warnings.
+
+    Parameters
+    ----------
+    data : (N, ...) array_like
+        Data to be z-scored
+    axis : int, optional
+        Axis to use to z-score data. Default: 0
+    ddof : int, optional
+        Delta degrees of freedom.  The divisor used in calculations is
+        `M - ddof`, where `M` is the number of elements along `axis` in
+        `comp`. Default: 1
+    comp : (M, ...) array_like
+        Distribution to z-score `data`. Should have same dimension as data
+        along `axis`. Default: `data`
+
+    Returns
+    -------
+    zarr : (N, ...) `numpy.ndarray`
+        Z-scored version of `data`
+    """
+
+    data = check_array(data, ensure_2d=False, allow_nd=True)
+
+    # check if z-score against another distribution or self
+    if comp is not None:
+        comp = check_array(comp, ensure_2d=False, allow_nd=True)
+    else:
+        comp = data
+
+    avg = comp.mean(axis=axis, keepdims=True)
+    stdev = comp.std(axis=axis, ddof=ddof, keepdims=True)
+    zeros = stdev == 0
+    # avoid DivideByZero errors
+    if np.any(zeros):
+        avg[zeros] = 0
+        stdev[zeros] = 1
+
+    zarr = (data - avg) / stdev
+    zarr[np.repeat(zeros, zarr.shape[axis], axis=axis)] = 0
+
+    return zarr
+
+
+def normalize(X, axis=0):
+    """
+    Normalizes `X` along `axis`
+
+    Utilizes Frobenius norm (or Hilbert-Schmidt norm / `L_{p,q}` norm where
+    `p=q=2`)
+
+    Parameters
+    ----------
+    X : (S, B) array_like
+        Input array
+    axis : int, optional
+        Axis for normalization. Default: 0
+
+    Returns
+    -------
+    normed : (S, B) `numpy.ndarray`
+        Normalized `X`
+    """
+
+    normed = np.array(X)
+    normal_base = np.linalg.norm(normed, axis=axis, keepdims=True)
+    # avoid DivideByZero errors
+    zero_items = np.where(normal_base == 0)
+    normal_base[zero_items] = 1
+    # normalize and re-set zero_items to 0
+    normed = normed / normal_base
+    normed[zero_items] = 0
+
+    return normed
 
 
 def rescale_test(X_train, X_test, Y_train, U, V):
@@ -24,7 +138,7 @@ def rescale_test(X_train, X_test, Y_train, U, V):
         Behavioral matrix, where `S2` is observations and `T` is features
     """
 
-    X_resc = utils.zscore(X_test, comp=X_train, axis=0, ddof=1)
+    X_resc = zscore(X_test, comp=X_train, axis=0, ddof=1)
     Y_pred = (X_resc @ U @ V.T) + Y_train.mean(axis=0, keepdims=True)
 
     return Y_pred
