+++ pyls/tests/test_utils.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 import numpy as np
-import pytest
 import pyls
 
 rs = np.random.RandomState(1234)
@@ -15,41 +14,6 @@ def test_RefDict():
     assert str(d) == 'ResDict()'
     assert(str(BlahDict(test={})) == 'BlahDict()')
     assert(str(BlahDict(test=10)) == 'BlahDict(test)')
-
-
-def test_zscore():
-    out = pyls.utils.zscore([[1]] * 10)
-    assert np.allclose(out, 0)
-
-    out = pyls.utils.zscore(rs.rand(10, 10))
-    assert out.shape == (10, 10)
-    assert not np.allclose(out, 0)
-
-
-def test_normalize():
-    X = rs.rand(10, 10)
-    out = pyls.utils.normalize(X, axis=0)
-    assert np.allclose(np.sum(out**2, axis=0), 1)
-
-    out = pyls.utils.normalize(X, axis=1)
-    assert np.allclose(np.sum(out**2, axis=1), 1)
-
-
-def test_xcorr():
-    X = rs.rand(20, 200)
-    Y = rs.rand(20, 25)
-
-    xcorr = pyls.utils.xcorr(X, Y)
-    assert xcorr.shape == (25, 200)
-    xcorr = pyls.utils.xcorr(X, Y, norm=False)
-    assert xcorr.shape == (25, 200)
-
-    with pytest.raises(ValueError):
-        pyls.utils.xcorr(X[:, 0], Y)
-    with pytest.raises(ValueError):
-        pyls.utils.xcorr(X[:, 0], Y[:, 0])
-    with pytest.raises(ValueError):
-        pyls.utils.xcorr(X[0:10], Y)
 
 
 def test_dummycode():
